Fetch Publication ID
# What

Retrieves a publication ID from a user

# Why
TL;DR: You need the publication Id to post to your blog's subdomain.

There are 2 ways to submit a post to HashNode. They are createStory and createPublicationStory. If you use createStory then it will post directly to HashNode and NOT under your subdomain. If you want to post to your subdomain you must post to createPublicationStory. To do so, you need your publicationId.